+++ web/packages/pyxis-user-site/vite.config.ts
@@ -18,6 +18,11 @@ export default defineConfig({
   server: {
     port: 3000,
     open: true,
+    proxy: {
+      '/api': 'http://localhost:8080',
+      '/auth': 'http://localhost:8080',
+      '/flyers': 'http://localhost:8080',
+    },
   },
   preview: {
     port: 3000,